Add pre-click AJAX path request loading
We now support pre-loading path AJAX requests when we think someone's about to
start them. When someone starts a click on a folder or nav breadcrumb via
`:on-mouse-down`, we now immediately send the request, then finish it once they
complete the click.

This will save anywhere from 50ms-100ms per click, depending on how fast a
clicker the user is. Locally, this is more than enough time to complete the
request entirely, so navigation becomes instantaneous.

This will have to be slightly modified to support touch devices (probably by
using `:on-touch-start` to handle pre-clicks), and should also eventually be
modified to be completely general such that any AJAX request could be
preflighted, once we're making different kinds of AJAX requests.